server/workers/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
google-cloud-storage==2.14.0
beautifulsoup4==4.12.2
lxml==4.9.3
asyncio==3.4.3
